+++ R/score_fp.R
@@ -14,9 +14,10 @@
 #' @importFrom magrittr %>%
 #' @export
 #' @author Yuanlong Hu
+#' @references Ye, H., Tang, K., Yang, L., Cao, Z., & Li, Y. (2012). Study of drug function based on similarity of pathway fingerprint. Protein & cell, 3(2), 132-139.
 #' @examples
 #'
-#'   data("drugSample")
+#'   data("drugResult")
 #'   res <- score_fp(drugResult$demoFP, n=100)