+++ NEWS.md
@@ -1,3 +1,10 @@
+CHANGES IN VERSION 0.9.0
+------------------------
+
+* Add network visualization function
+* Add rank aggregation methods
+
+
 CHANGES IN VERSION 0.8.4
 ------------------------